(chore) replace source-string ui tests with render and contract checks
