+++ tests/config/arena-drain-config.test.ts
@@ -9,8 +9,6 @@ import {
 const vercelConfig = JSON.parse(readFileSync("vercel.json", "utf8")) as {
   crons?: Array<{ path?: string; schedule?: string }>;
 };
-const voteRouteSource = readFileSync("app/api/arena/vote/route.ts", "utf8");
-const voteJobsSource = readFileSync("lib/arena/voteJobs.ts", "utf8");
 
 const voteDefaults = resolveArenaDrainRequestLimits(
   new URL("https://minebench.ai/api/admin/arena/drain-vote-jobs"),
@@ -59,22 +57,6 @@ assert.deepEqual(
 assert.equal(shouldScheduleArenaVoteJobDrainAfterResponse(1, undefined), true);
 assert.equal(shouldScheduleArenaVoteJobDrainAfterResponse(4, "0"), false);
 assert.equal(shouldScheduleArenaVoteJobDrainAfterResponse(0, "1"), false);
-assert.ok(
-  voteRouteSource.includes("recordArenaVoteQueuedForSampling"),
-  "vote route should immediately update sampling coverage for queued decisive votes",
-);
-assert.ok(
-  voteRouteSource.includes("scheduleArenaVoteJobDrain()"),
-  "vote route should schedule the bounded vote job drainer after successful queue writes",
-);
-assert.ok(
-  voteJobsSource.includes("recordArenaVoteInSamplingCache"),
-  "vote drains should update the cached sampling state instead of clearing it",
-);
-assert.ok(
-  !voteJobsSource.includes("invalidateArenaCoverageCache"),
-  "vote drains should not invalidate the 60s sampling cache after every processed vote",
-);
 
 assert.equal(
   shouldIncludeArenaDrainStatus(
